MB-24572: Stop Collection aware DCP from sending empty snapshots
If a DCP stream is configured with filtering enabled and the filter
results in 0 mutations, we still send snapshot markers to the client.
In a needle/haystack situation, the client looking for needles will
get frequent traffic (snapshot markers) triggered by the mutations
made against the haystack.

With this change, filtering is moved earlier in the stream flow, now
when we build the 'readyQueue' we apply filtering allowing the
existing empty checkpoint/snapshot logic to prevent an empty snapshot
being sent.

Change-Id: I3379a9ea4ff11aed4f287f6cb688c8af0eecd8db