MB-25342: Fix missing completeDeletion when collection is reopend
When updating the CollectionsFlushTest to use the real collections
erase flow (i.e. compaction) the test failed because the final phase
does a sequence of

1. create fruit
2. delete fruit
3. create fruit
4. trigger compaction/erase

So when the compactor/eraser processes the bySeqno index it finds all
of the logically deleted fruit keys, but never triggered 'complete'
because the final 'fruit' system key is a create marker for the
new generation of fruit. That key was not logicallyDeleted so we
just returned false from KVBucket::collectionsEraseKey at line 2407.

We now don't shortcut out so early, but proceed to test if the key
is a trigger (shouldCompleteDeletion => true) for completeDeletion.

This commit was previously a simple test update, but that found the
issue so now this commit does both.

Change-Id: Id8298da4fef33ae098d2ebc9063481a570f09d61