MB-27457: [5/n] Allow KV-engine to set the delete time
1. Update couch-kvstore so that when a delete is persisted we check
if it already has a time assigned to it, resulting in no call to get
the current time.

2. Update DCP consumer so that a V2 delete interface is available
enabling the consumer to accept a deleteTime and set it into the item.

3. No change required for DeleteWithMeta, previously all exp were
0 (certainly for XDCR) so couch-kvstore just picks up the incoming
value.

Tests are added which demonstrates that the new DCP API can be used
to set override the deleteTime and same for DeleteWithMeta.

Change-Id: Id79ec006ee37a62d740827054b8b369933dc3253