MB-28453: [1/n] Switch xattrs from byte_buffer to char_buffer
Overall we wish to remove the use of byte_buffer as we frequently
end up casting it to char. This commit just switches xattr code from
using a byte_buffer to char_buffer, making it easier to add
decompression support to xattr::Blob which is required to cleanly
address the MB this commit is linked with.

Change-Id: I197723539ea1f7c799ccf3dcddfa1757531e36ba