MB-: Collections by-ID conversion
Primary change is that kv-engine polices collections
using the collection ID, not a prefix on the key.

DocNamespace is now CollectionID (although to reduce the
patch, they are both the same and will be unified later).
The CollectionID is now a uint32 rather than an enum and
serves a similar purpose in-side KV as to what the
DocNamespace used to, that is KV includes the ID in hashing
and as a prefix in the disk indexes. We reserve 0 and 1
for our own use, but other values can be assigned to
collections.

The Manifest still knows about collection names, but once
collections are pushed down to the vbucket, only the ID
is known about for example DCP system events only transmit
the ID and the _local JSON manifest stores IDs only.

KV operations must now encode the desired collection ID
which is encoded in the key-bytes (if HELLO collections).
Clients which don't enable collections just get their
operations assigned the DefaultCollection ID (as has
always been the case).

Some objects with KV now increase in size because the
namespace is no longer a byte, but a uint32, StoredDocKey
and SerialisedDocKey grow in size, and they are members
of Item and StoredValue. However users no longer need
to encode a prefix and separator in the key-name.

Change-Id: I583dd0e12b4b17057d325939a2da60ef16ff78b2