MB-31978: DcpProducer map vb to a container of streams
In preparation for storing more than one stream object per VB, change the
DcpProducer map type so that it maps to a container.

AtomicUnorderedMap is given new methods and tests which will be used
in this patch and in later patches of this feature.

In this patch the container is a std::vector and we only ever store a single
stream, but we update code that is dependent on the stream map value to work
with a container of many streams.

Change-Id: If9f2f43d0ee9dfb560a4276ecd0e760e4b3aa773