MB-31978: Add StreamContainer and unit tests
StreamContainer is a custom container designed for use by the
DcpProducer so that it can enable the multiple streams per VB feature.
The class is templated primarily to simplify testing and is not
attempting to be a generic container.

A StreamContainer default constructs to store a single element, as that
is how the DcpProducer will initialise a vbucket.

A StreamContainer exposes a thread-safe API and internally uses shared
locking to enable multiple readers. A read API is exposed via a
ReadLockedHandle obtained by calling lock() and a  write API is exposed
via WriteLockedHandle obtained by calling wlock().

The ReadLockedHandle and WriteLockedHandle both also expose iteration
methods begin()/end()/next() so that the elements of the container can
be iterated.

To support the DcpProducer a ResumableIterationHandle can be obtained by
calling startResumable(). A ResumableIterationHandle is an object which
provides an iteration style interface, but crucially remembers the
current position, this allows the user to begin iterating the elements
and destruct their ResumableIterationHandle and later resume from where
they had previously iterated to (providing that the StreamContainer
membership hasn't changed).

Examples of ResumableIterationHandle

If the container store integers 5, 4, 3, 2 and 1 an initial loop (begin
to end) yields:

   54321

A new ResumableIterationHandle will from 5, but assume this time we
break the loop 3 is returned:

   543 <break>

The next ResumableIterationHandle will start from 2 and will end when a
complete cycle of the elements occurred:

Now a new loop will start from 2 and would return true from complete
when a full cycle has completed, e.g.

   21543

Resuming iteration this time again starts from 2:

   21543

Note1: only a single resume point is maintained, so if two threads were
creating ResumableIterationHandle from the same StreamContainer, they
will be interfering with each others resume point.

Note2: Inserting elements or removing them from the StreamContainer
resets the resume point.

Change-Id: Ie7a6d6bc9fda9ec610a9cd09519ffaba825b5e33