MB-31266: Add system-events to ephemeral seq index linked-list
Persistent buckets keep the system events on disk in the seqno index
and ephemeral needs equivalent functionality.

This commit refactors VBucket::queueItem so we have a persistent vs
ephemeral implementation (and renames it).

In ephemeral we store the system event in the hash-table which in
turn then weaves that OrderedStoredValue into the seqno linked list.

Testing is update to port a number of DCP tests to run for all bucket
types, any tests with explicit warmup are fixed up in a future patch.

Change-Id: Iccae24c21fdc426dbe4ba4fcd99cbae5c6492969