MB-32650: Change collections feature code from 0x9 to 0x12
Older kv-engine releases are unconditionally granting the collections
feature code (of 0x9) meaning the 3.0 SDK (Which will support
collections)can think collections are supported when talking to say
Couchbase 6.0 but then finds the 6.0 server will generate errors when
the client  attempts to use the feature (as the wire protocol is
different).

To avoid this situation, collections changes to feature code 0x12 and
the SDK should follow, now when we GA the server and 3.0 SDK, we will
never have any supported code requesting feature 0x9 against any CB
version.

Change-Id: If553f9a1e0dfae29af50cc5177a63f5a5c263a5d