MB-32784: 2/5 Collect collections meta data for persistence
Note this patch maintains two concurrent collection's meta-data
mechanisms. The old and the new... The new code is generating and
reading back persistent meta-data through new KVStore methods (tested in
a new GTEST suite) and the old mechanism is still used and tested by the
rest of the collections code and tests.

The overall plan to address MB-32784 is to give KVStore ownership of the
persisted collection's data, replacing the current mechanism where the
persisted data is controlled from the VB::Manifest.

As the flusher writes systems events (setSystemEvent and delSystemEvent)
the following structures track the collection state by inspecting the
system event Items.

1) A vector of queued_item's that represent collection creations. The
   code captures the whole Item to simplify the meta data capture
   (rather than mangling from one structure to another).
2) A vector of collections that have been deleted. This object captures
   the ID and end-seqno of each dropped collection.
3) A vector of scopes (just the ID) that have been created
4) A vector of scopes (just the ID) that have been dropped.
5) The highest manifest-UID seen in the flush.

Now when we commit the flush, the final implementation is expected to
persist this data in any suitable form.

For couch-kvstore we choose to persist the meta data to 4 '_local'
documents as flatbuffers tables (schema -> collections/kvstore.fbs).

"_local/collections/manifest" stores the manifest-UID, this is an open
and write (no read of existing value).

The rest of the update is performed by a read, update and write.

"_local/collections/open" stores a vector of collections that are open.
To update the open collections is a read of the existing doc occurs, the
open collections read from persistence are checked against the vector of
dropped collections and the output is merged with the vector of newly
created collections, the final result is written back as the new set of
open collection.

"_local/collections/dropped" stores a vector of dropped collections.
The requirement to store a vector of dropped collections depends on the
kvstore's implementations ability to atomically drop the entire
collection. As couchstore cannot discard an entire collection as part of
the commit, it maintains a list collections that the compactor will
later be able to use for dropping keys. couch-kvstore maintains the
document by reading the existing document and merging in the new
collections and writing the result.

"_local/collections/scopes" stores a vector of known scopes, this
document is maintained by a read of any existing document and then a
merge of the scopes which were created and dropped in the flush.

Finally KVStore must be able to retrieve via getCollectionsManifest_new
 - this API name will eventually replace getCollectionsManifest

1) The manifest-UID
2) The set of open collections
3) The set of open scopes

Change-Id: If29ee48ba7d7faf50feebed17659800d6dc8b68a