MB-23086: Use cb::ArenaMalloc to give each EP bucket an arena
Utilise the cb::ArenaAlloc library to register an arena for each EP
bucket (persistent/ephemeral). All heap alloc/dealloc should be
tracked in the registered arena, to achieve this the existing
onSwitchThread will also switch to the engine's arena.

Memory tracking is now changed to just ask cb::ArenaAlloc how many
bytes are allocated to a given arena - however there are now some
changes.

getEstimatedTotalMemoryUsed:

This EPStats method is effectively unchanged, it just returns the
value of a single atomic counter and can be ahead or behind a more
accurate counter. However that value will now remain permanently
'stale', which can be an issue for tests. Hot code patihs continue to
use getEstimatedTotalMemoryUsed

The reason for the staleness is that platform no longer calls the
new/delete hooks which would of triggered occasional updates to the
estimated counter, instead tests must do this themselves or be-aware
that in the full stack two events periodically trigger updates.

getPreciseTotalMemoryUsed:

This EPStats method is changed and now asks cb::ArenaAlloc for the
current total the engine's arena has allocated. A side affect calling
getPreciseTotalMemoryUsed function is an update to the
getEstimatedTotalMemoryUsed counter.

Test code must now either call getPreciseTotalBytesUsed in expect
statements and/or put explicit calls to getPreciseTotalBytesUsed ahead
of entering engine code where the test desires some memory usage
behaviour. The majority of the changes in this commit are test changes
of this nature.

Above it is mentioned that in the 'full-stack' two events will trigger
updates to the getEstimatedTotalMemoryUsed, these are

1) The period request from ns_server for engine memory stats, which
uses getPreciseTotalMemoryUsed to set the "bytes" and "mem_used"
stats.

2) The MemoryTracker thread calling
cb::ArenaMalloc::updateTotalCounters every time the task wakes up.

Note: This patch explicitly leaves dead code and a split of
responsibility between MemoryTracker (and children) and the new
platform code. This is to keep the changes in this patch as functional
as possible, subsequent patches will perform removal and re-factoring.

Change-Id: Idec726ea534107d0b4d7aebf9f57d2ef1b7f3c84