MB-23086: Remove memory allocator hooks
Remove the AllocApi memory allocator hook functions as cb::ArenaMalloc
now tracks without hooks.

As part of the removal some test code is now moved from kv_engine to
platform.

With this commit there is a reduction in 'introspection' with those
benchmark tests that exploited the hooks API to track memory peak during
a workload.

Change-Id: I49708378f1afa83c9dcd35a8ba16535eac9c91b8