MB-23086: Switch engine TLS to null after running a task
If an engine pointer is left in the ObjectRegistry Thread Local
Storage after the ExecutorThread has completed a task, the pointer can
become invalid if the bucket is shutdown, this isn't yet a serious
problem as the pointer is never dereferenced (except for one debug
only log message). However in the main patch for MB-23086
onSwitchThread will need to dereference the TLS engine to obtain an
'arena' object.

This patch also includes a cleanup of the stopTaskGroup call.

1) We should ask for every task to be stopped, not just nonio
2) We no longer need explicit cancels after calling stopTaskGroup

Change-Id: I6fd9d4b4f2ce1c8df0351ffd3c315286f4698624