MB-23086: Map a bucket to a JeMalloc arena with cb::ArenaMalloc
This patch makes use of the new cb::ArenaMalloc (platform) API
to associate a bucket with its own JeMalloc arena, this will
give the ability to view memory allocation statistics on a per
bucket rather than per process basis.

cb::ArenaMalloc requires that

1) A bucket (client) registers for an arena and de-registers when done

The existing create_ep_engine_instance code now registers for an arena
before allocating the EventuallyPersistentEngine object (so we now
account for the sizeof  EventuallyPersistentEngine and all allocations
done in construction in memory tracking). The engine's arena is given to
the engine via the constructor and held onto for further interaction
with cb::ArenaMalloc

2) An executing thread must tell cb::ArenaMalloc which client is now
   executing (the same pattern as ObjectRegister::onSwitchThread) and
   when finished.

Since we already have the ObjectRegistry::onSwitchThread code/pattern,
that code is modified to also tell cb::ArenaMalloc which client is
executing.

Finally cb::ArenaMalloc integrates memory tracking into the API, the
memory tracking hooks are now never executed by cb_malloc.

The existing memory used precise and estimate are supported by
cb::ArenaMalloc so the EPStats methods now just call through to ask the
Arena.

There are a few other changes where we must configure per core memory
update thresholds, again this code already exists, it is now just
redirected to call cb::ArenaMalloc.

Note:

* This patch explicitly leaves dead code (alloc_hooks/MemoryTracker).
This is to keep the changes in this patch as functional as possible,
subsequent patches will perform removal and re-factoring.

* allocator stats needs work.

Change-Id: Ida6d3d60b8af5cd85f0af8443f2efd720edee2e5