MB-36948: End non-sync write backfills on a visible item
Some DCP consumers need to receive the last item in a snapshot to
confirm that it is complete (e.g., indexer).

If these consumers are not sync write aware, they will not be sent
prepares or aborts (and commits will be replaced with mutations).

If such a consumer receives a snapshot from disk which ends on the
seqno of a prepare or abort, it will *not* receive the last item.
(Snapshots from memory do not encounter this issue, prepares and aborts
are filtered out before determining the snap start and end).

To resolve this, adjust the disk snapshot end to the most recent
committed item available from disk.

There are also edge cases where the entire backfill range may not
contain any items.

Change-Id: Ifbda704554215d774069f683f1426a33c5bf3bd6