MB-36948: ActiveStream track maxVisibleSeqno for in-memory snapshot
As the ActiveStream transfers items from the checkpoint to the
readyQueue, check if the item is committed, and if so update
a maxVisibleSeqno which will then be added to the SnapshotMarker
response message.

Updates a number of unit tests to inspect the snapshot-marker.
A number of those expects will change once the backfill passes
the maxVisibleSeq through.

Change-Id: I142565158f7c513ea98ec8677eb3511f656d0dbf