MB-36948: Track and store maxVisibleEndSeqno in Checkpoints
Add a maxVisibleEndSeqno to Checkpoint and then maintain that
value for active/replica.

Replica vbuckets set this value on receipt of a snapshot-marker.
They will either call createSnapshot or updateCurrentSnapshot both
methods now require the visible snap-end value. The value then
remains in the Checkpoint object and is now visible to subsequent
DCP producers (producing the replica) or readable from
CheckpointManager (which will be wired into GetAllVBSeqnos).

Active VBuckets track/set the maxVisibleEndSeqno in a way which
mirrors the visible seqno (which is CheckpointManager::lastBySeqno).
As items are queued (queueDirty) lastBySeqno is increased and
the current Checkpoint will have it's end seqno adjusted.
The same logic applies to a new member, maxVisibleSeqno
which only increments for visible items.

Note on tests using "createSnapshot": to avoid a default parameter
the createSnapshot method has been updated in many unit tests and
in some places what seems like the correct maxVisibleSeqno is
added and sometime later tested. Many tests though are just using
zero, as for many tests the maxVisibleSeqno isn't of interest.

Change-Id: I0e2a6aafeeef3a334766e8bdc868471462510d85