MB-37772: Refactor ScanContext
A number of re-factoring changes around ScanContext.

1) ScanContext is now RAII in that destruction of the object will
trigger destruction of the KVFileHandle (which e.g. on couchstore,
closes the database backing the scan).

With that:

1a) initScanContext now returns a unique_ptr
1b) scan expects a non const reference to a context
1c) destroyScanContext is now redundant and removed

2) ScanContext was constructed with shared pointers to the two callbacks
it needed, this is not required and references to callbacks are
sufficient.

2a) This required a small amount of object life-time tweaking in
    DCPBackfillDisk
2b) A knock on effect is that the rollback interface also must specify
   a callback of the correct type (reference).

3) ScanContext itself is given a hierarchy in preparation for
by-id scans. ScanContext is the base class (owning callback references
amongst other data). BySeqnoScanContext is a class which stores
information about the scan seqno range.

Change-Id: Id51a0a4bffddc10c3dd6bc0ad9d31e4bb5bfc61a