MB-38673: stats scopes - don't print scope name multiple times
When running the scopes stat, it would return the scope name and then
each collection was also returning the name of its owning scope. We
only need the collection stat code to print the name when it is
called for the 'collections' stat.

Also tidy up logging from previous refactoring

Change-Id: Ic111b587067deecb56927d21eec7d45580fc8b5d