MB-39594: Add constructors for non system-event code paths
Profiling a collection enabled workload showed functions taking
a higher percentage compared to a non-collection workload. This
MB aims to reduce the following:

* Item::Item called from
 * EventuallyPersistentEngine::itemAllocate
 * StoredValue::toItemBase

Item::Item is inspecting the key, "is the key in the system-event
namespace?" For these code-paths there is no need to inspect the
key, which requires a few loads/compares (more for longer IDs).

These front-end Item paths can use constructors which don't do
the key checks.

Change-Id: I37cee0468fdf84496d806e566e506a3945eb3fa5