MB-39594: Reduce work required for hash-table key compare
Key comparison in the hash-table uses SerialisedDocKey::hasKey (and
the operator==)

With collection-IDs being canonical the SerialisedDocKey::operator==
can be improved to remove the decoding of collection-IDs.

When the 'rhs' DocKey says it encodes a collection, the operator
can just compare size/data.

If the 'rhs' DocKey does not encode a collection we adjust the compare,
first look at the SerialisedDocKey byte 0, then we can compare
size/data.

Change-Id: Ib912851992ccae9170c67e46272750ec3403f726