MB-40258: Wrap dbFileRevMap with folly::Synchronized
Remove the CouchKVStore owned mutex and use folly::Synchronized for
access to the revision-map. This fixes MB-40258 by ensuring 1 lock
exists for use by both RW and RO instances of CouchKVStore.

This is a minimal patch which can be candidate for back-porting without
re-factoring.

Testing:

The CompactAndScan test was originally added to test for MB-27963, and
it does for the code added for MB-27963. However it doesn't account for
the mistake in MB-27963 where we have different instances of
CouchKVStore.

This commit updates CompactAndScan so that there are a read-only and
read-write instance of the store, allowing for the race to occur.

Note to see the race occur it did require an increase in iterations
(currently 10), this is not a change included in this patch as it took
an unreasonable amount of time to 'reliably' trigger the failure.

Change-Id: Icff820aa059dececffd752b821e6bc67e05ce823