MB-39864: Don't underflow collection item count after recreating a collection
When KV recreates a collection the following occurred which causes an
underflow.

1. create:c1
2. mutate(c1, key=x)

Now c1 has an item count of 1.

3. drop:c1
4. create:c1

Now c1 has an item count of 0, it's a 'new' collection

5. mutate(c1, key=x)

At this point, KV got the item count wrong because on disk a copy of
key=x exists from step 2. The stat update code treats the insert as
an update and does not increment the item count.

6. delete(c1, key=x)

Here we now underflow as the delete of key=x decrements the item count,
which is currently 0.

The fix to this issue is to use the collection's "isLogicallyDeleted"
function that takes a key and a seqno and determines if the key belongs
to a dropped collection or exists in the range of seqnos for an old
generation of a collection.

With this fix in place, step 5. now treats the existing key=x as
deleted, and 5. is correctly treated as an insert, raising the item
count of collection c1 to 1, step 6 no longer underflows.

Change-Id: I6bbba795579fbf97348186ea0ab7af1bbde8db90