MB-39864: Clean-up dropped collection meta-data
Change how we detect the need to remove collection's from
the set of dropped IDs. With collection resurrection (or indeed
collection flush) a collection can be dropped yet have a non
deleted system-event. When seqno indexed purging hits any collection
system-event we unconditionally check for the ID in the set
of dropped collections.

Change-Id: I04e691cb279d8e5c6d9d957e72ad03b55c58ab5d