MB-39864: Collection stat accounting when a collection is resurrected
The primary issue addressed in this patch is to ensure that when
we do collection stat updates, we ignore any existing item if the
item belongs to a dropped collection. This ensures we don't treat
an insert as an update and get the collection stat wrong (which
can lead to underflow of the collection item counter).

The commit moves all collection stat updated code into a pair of
functions that decouples the update code from our kvstore
implementations.

Within these new functions a new check is made, asking if the
old version of the flushed key is logically deleted, switching
updates to inserts.

Change-Id: I53c0853dcac12de60879e957fb14a06162800e2a