MB-41321: Defer collection statistic updates until successful commit
Collection statistics for persistent buckets are updated from the
flusher before we commit, the reason this is done like that is so
we can do '_local' updates to include collection statistic changes made
in the flusher batch.

MB-41321 identified that this approach was flawed and can result in
incorrect statistics or underflow exceptions when commit fails. When
the flusher fails, the current batch of items is retried, this means
each retry will apply the same changes to the collection statistics,
we will then count items again for each flush retry.

This commit changes how collection statistics are updated to occur in
multiple steps.

1) As the items of the batch are processed, we now update a flusher
owned map of collection-ID to collection statistics - this collects
the 'deltas' of changes being made by the flusher batch.

2) Before commit we read the current collection statistics and apply
the collected changes to generate statistics for the '_local' updates.

3) If the commit is successful, we update the current in-memory
statistics.

If the flush is retried, the flusher owned 'collector' is recreated
for each retry, so each retry just counts for its run.

Change-Id: Icf57856233843646111e63126bed95634e169412